refactor: extract tags mod for dataset.rs
